Support falling back to space separated charset
In some cases, for example, Apple Mail, programs generate HTML
with apallingly bad meta tags such as:

    <meta content="text/html charset=utf-8">

This is bad because *a* no http-equiv="Content-Type" and *b* because
the content type and charset do not have a separating semi-colon.

Sadly, Chrome et-al support this, so we need to in Hubbub.  This
change adjusts the content="" parser to retry if it cannot find
a semicolon, and work forwards to first whitespace instead.

Fixes: #2549